refactor: split parts data modules
