+++ lib/data/parts-server.ts
@@ -1,13 +1,6 @@
 import 'server-only'
 
-export {
-  searchParts,
-  getPartById,
-  getPartDetail,
-  createEmergingPart,
-  updatePart,
-  getPartRelationships,
-  getPartNotes,
-  logRelationship,
-} from './parts'
+export { searchParts, getPartById, getPartDetail, getPartNotes } from './parts.read'
+export { createEmergingPart, updatePart } from './parts.mutate'
+export { getPartRelationships, logRelationship } from './parts.relationship'
